// frontend/src/app/models/infoshop.ts
import {Image} from "./image";

export class Infoshop{
  id: number;
  name: string;
  logo: Image[];
  description: string;
  introduce: string;
  phone: string;
  gmail: string;
  address: string;
  linkMessage: string;
  linkGoogleMap: string;
  imgDescription: Image[];
  linkFanpage: string;
  linkShoppe: string;
  linkMap: string;
  linkZalo: string;
}